Калькулятор

+++ src/main/java/faang/school/godbless/BJS2_796/Calculator.java
@@ -1,6 +1,8 @@
 package faang.school.godbless.BJS2_796;
 
+import java.math.BigInteger;
+
 @FunctionalInterface
 public interface Calculator {
-    long operate (int a, int b);
+    long operate(long a, long b);
 }
